Split up signed_log_test, fixed persistent_bw_test
The signed_log_test.cpp file was getting too long and unwieldy, and most
of the methods couldn't be written inline in the class definitions
anyway, so I split it into a header and source file.

Now that I know how to use a shared test state object to get the
persistent version of the last message in the experiment, I can apply
the same setup to persistent_bw_test, which also relied on the old
global stability callback. This test is much simpler, so I'll try a
TestState object with no methods, and leave the state-updating logic in
the RPC method or persistence callback.
